Removed the NSURLSessionDownloadDeletegate code from AppDelegate.  Simply the wrong place.
Moved the mpPlayerLoadStateDidChange method to the MyTableViewController as player is setup in viewDidAppear

Moved all of the Globals and defaults initialization calls from didFinishLaunching to the method loadSermons() in MyTableViewController where it is used to initially load the sermons and defaults as well as in handleRefresh so the behaviors are the same.

Moved setupSermonPlaying method from AppDelegate to MyTableViewController as noted above.

Fixed hasNotes(check) & checkNotes and hasSlides(check) & checkSlides to make them actually use the check boolean passed and to not do a check if Reachability isn't true.