Got rid of didRotateFromInterfaceOrientation and used vieWillTranstionToSize
Implemented setupSplitViewController() throughout to properly control when the master view controller appears and that it doesn't for full screen webViewControllers.  This meant putting it in viewWill or viewDid Appear and also in viewWillTransitionToSize for TVC, MVC, and WVC.

Fixed a bug in setupSermonsInSeries where it wasn't include the sermon itself, which it always should, regardless of whether that sermon is in a series proper.  Now it does.