Restructured the Globals to put a lot of the related variables into structs.
Changed the cache so it is cumulative during the user session.  That avoids recreating entries which can save a lot of time for frequently used ones.

Still have to clear and rebuild the cache when doing a refresh, i.e. downloading JSON and rebuilding the sermon array.

Fixed bug where sermonsInSeries list wasn't hidden on refresh.

Fixed an arcane bug related to sermonYears.