Code clean-up
Clarified download to refer to audio

Made sure UIBarButtonItems (top left and toolbar) in TVC were disabled at the proper times (during startup and refresh) and that the calls to disable or enable were in one function, not directly in the code.

Made sure the sermon list disappeared during refresh

Fixed bug in setAudioOrVideo where captureContentOffsetAndZoomScale() was only being called if the sermon selected was the sermon currently playing when it should have been (and now is) called every time.

Added settings struct to Sermon to front the Globals.sermonSettings dictionary of dictionaries, allocate the dictionaries (i.e. a dictionary for each sermon) when needed,  and make the code cleaner.