Found out how to use scrollView delegates within a WKWebView - you must set the delegate to nil in viewWillDisappear to avoid crashing.  Now content offset and zooming can be captured immediately after they happen.
Called setContentOffset and setZoomScale only from the main queue, that seems to be required to avoid problems.

Changed MyViewController to not do anything when the same row in a series is selected again.

Added saveSermonSettingsBackground and used that widely.

Restructured loadSermons()

Addressed when refreshControl would be active to avoid it being activated during startup sermon loading.