Massive bug fixing in MyViewController.  Needed to use TWO (2) NSTimers, one of reach WKWebView for loading.
Also needed to get the processing right to differentiate the downloading/loading from what is showing, i.e. UI.

Also needed separate downloading/loading selectors for each of the notes and slides WKWebView.

Added Check For Update capability to allow cached documents to be re-downloaded individually to get updates on the server.