Moved to a generalized documents model rather than just notes and slides.  Now there can be any number of documents.
Each document has its own timer for loading, as a single timer was simply wrong and creating problems.

The "purpose" of a document is the same as "showing" for what is visible.

Changed Sorting strings to be Chrono to Newest to Oldest, and vice versa, to be more user friendly.

The strings presented in the user have the same order, number, and meaning as the internal strings used but do not have to be the same.

Changed to popover tableview for About action menu.