Removed pause in background when video is playing.  Leaving that to the system.
Added ability to recover to known good JSON in Bundle when bad JSON is downloaded.

Fixed bug in MVC viewWillDisappear.  The  scrollView delegates of ALL documents must be set to nil on exit.  Same thing must be done in didSet for selectedSermon when value is nil.

Finally fixed the problem w/ iPhone live view - errant UINavigationController in front of the Live View!