Lots of bug chasing around UIKit changes from background thread after it was changed from the main thread, and also on KVO AVPlayerItem.status and removeObserver causing a crash because the object it was being removed from wasn't being observed.
Didn't solve either one, put in a lot of NSThread.isMainThread guard statements, but don't have a clue on either one.  Don't even know if the removeObserver crash is coming from the global or the MVC removeObserver, or both.

Fixed a bug where the PIP player play/pause was changed and sliderTimer() wasn't picking it up.